wip
- change enum Board_type to uint32_t
- add zeyu_hand esc id

+++ include/iit/ecat/advr/foot_sensor_esc.h
@@ -145,7 +145,7 @@ public:
         return SDOs;
     }
     virtual void init_SDOs ( void );
-    virtual uint16_t get_ESC_type() {
+    virtual uint32_t get_ESC_type() {
         return FOOT_SENSOR;
     }
